GH_ISSUE_923: split objects_read.go into focused read-path files
736-line objects_read.go mixed Manager construction, pre-flight
checks, GET orchestration, HEAD, LIST, HTTP-range parsing, and the
streaming-SHA-256 verifying reader. Pure structural move; no logic
changes.

  - manager.go: Manager type, Deps, New, invalidateObjectCaches,
    LocationCache, managerSpanPrefix, plus the pre-flight surface
    (CanAcceptWrite, BackendCapacityStats, ListObjectsMaxPages,
    ObjectExists). The read-side skeleton.
  - get.go: GetObject, tryGetObjectCache, getAttemptRequest,
    getObjectAttempt, resolveBackendRange, maybeWrapIntegrityReader,
    populateObjectCache.
  - head.go: HeadObject.
  - list.go: AdvancePastEmittedCommonPrefix, ListObjectsV2Result,
    ListObjects, listObjectsError, consumeListPage, tryEmitCommonPrefix.
  - range.go: ParsePlaintextRange.
  - integrity_reader.go: HashBody, VerifyingReader and methods.

cache_populate.go was deliberately skipped - the two cache helpers
(tryGetObjectCache, populateObjectCache) live with the GET flow they
serve.
