+++ .version
@@ -1,1 +1,1 @@
-v0.60.10
+v0.60.11
